requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
pydantic==2.10.3
markdown==3.7
pyyaml==6.0.2
